Add template for bills
The bill template places the content of a esr paying-in slip at the bottom of the page.
This is ment to be printed on a paper with the paying-in slip preprinted on it.
For the esr the package from Christian Luginbühl <dinkel@pimprecords.com> is used.
To prevent the need for installing fonts, the esr package was slitly modified to use the ocr
package which prowides the needed ocrb font